requirements.txt
iniconfig==2.0.0
packaging==24.0
pip3-autoremove==1.2.2
pluggy==1.5.0
psycopg2-binary==2.9.9
pytest==8.2.2
setuptools==70.0.0
